Data soon to be ready for Random Forest
I have implemented a better system for retrieving stock data by adding a "$" before all stocks. Since it is almost certain that anything in r/wallstreetbets that is "$" + "Stock abbreviation" is a real stock and not a false positive, this is a reliable way to get data that we can be sure is referencing the stocks we want. However, it VASTLY reduces the available data, by almost a factor of ten.

As far other work, I have implemented a few textual analysis features that will be quite useful. This includes adding text length, sentiment analysis, reading level, and other features of the text. These will help feed our random forest model.

TODO: Add the API data from Yahoo Finance, and start getting the DELTA (net change in share price) to get the final piece of the data.